Bump vimeo/psalm from 6.1.0 to 6.8.8 and fix new issues
Bumps the all-composer group with 1 update: [vimeo/psalm](https://github.com/vimeo/psalm).

Updates `vimeo/psalm` from 6.1.0 to 6.8.8
- [Release notes](https://github.com/vimeo/psalm/releases)
- [Commits](https://github.com/vimeo/psalm/compare/6.1.0...6.8.8)

---
updated-dependencies:
- dependency-name: vimeo/psalm
  dependency-type: direct:development
  update-type: version-update:semver-minor
  dependency-group: all-composer
...

+++ functions/buildImageList.php
@@ -121,7 +121,7 @@ function buildImageList ( $unsafe_currDir, $maxColumns, $maxRows,
     $max_row = $maxRows + 1;
     $firstThumb = $mig_config['startfrom'] * $max_col * $max_row;
     // This rounds off any fractional part
-    $pages = ceil($thumbsInFolder / ($max_col * $max_row));
+    $pages = (int) ceil($thumbsInFolder / ($max_col * $max_row));
 
     // show last page if $startfrom to big
     if ($thumbsInFolder<$firstThumb) {
